% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_betadiv.R
\name{amp_betadiv}
\alias{amp_betadiv}
\title{Beta-diversity plot}
\usage{
amp_betadiv(data, distmeasure = "bray", label_by = NULL,
  show_values = TRUE, values_size = 3)
}
\arguments{
\item{values_size}{}
}
\description{
Beta-diversity plot
}
\examples{
amp_betadiv(AalborgWWTPs, show_values = FALSE)
}
