% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_mergereplicates.R
\name{amp_mergereplicates}
\alias{amp_mergereplicates}
\title{Merge replicate samples}
\usage{
amp_mergereplicates(data, merge_var, merge_fun = "mean", round = "up")
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{merge_var}{Variable in the sample metadata that defines the sample groups}

\item{merge_fun}{Name of the function used to merge the samples, fx \code{mean} or \code{sum}. (\emph{default:} \code{"mean"})}

\item{round}{Read counts are whole numbers, so if any decimals after merging they must be rounded either \code{"up"} or \code{"down"}. (\emph{default:} \code{"up"})}
}
\value{
An object of class \code{ampvis2}
}
\description{
Merge replicate samples
}
\examples{
library(ampvis2extras)
d <- amp_load(ampvis2::example_otutable, ampvis2::example_metadata)
d$metadata
d$metadata$group <- c("group1", "group1", "group2", "group2", "group2", "group3", "group4", "group4")
d$metadata
dmerged <- amp_mergereplicates(d,
  merge_var = "group",
  merge_fun = "mean",
  round = "up"
)
dmerged
}
