% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amp_betadiv.R
\name{amp_betadiv}
\alias{amp_betadiv}
\title{Beta-diversity plot}
\usage{
amp_betadiv(data, distmeasure = "bray", label_by = NULL,
  show_values = TRUE, values_size = 3)
}
\arguments{
\item{data}{(\emph{required}) Data list as loaded with \code{amp_load()}.}

\item{distmeasure}{Distance measure}

\item{label_by}{label}

\item{show_values}{\code{TRUE}/\code{FALSE}}

\item{values_size}{size}
}
\value{
A ggplot2 object.
}
\description{
Beta-diversity plot
}
\examples{
amp_betadiv(AalborgWWTPs, show_values = FALSE)
}
